:blowfish:Reimplement terrain map streaming mechanism
We have moved the terrain generation from real-time streaming, i.e., send chunk to renderer as soon as one is done, to the way that returns only when all chunks in a render distance are available. This is done because `STPWorldPipeline` already has a double buffer system that is swapped only when all generations are done, so it is no longer necessary to stream chunks immediately to back buffer.

Generator manager in `STPWorldPipeline` has been revised with added implementation for efficient neighbour condition checking and scheduling to avoid race condition.

# STPChunk
- Change the function *calcChunkNeighbour* to *calcChunkNeighbourOffset*. The difference is the prior one returns chunk world coordinate and the latter one returns relative offset, which can be converted to world coordinate by adding the coordinate of the centre chunk. This can help the world pipeline to cache neighbour offsets and reuse to avoid repetitive small memory allocation.
- Remove all internal states tracking such as occupancy status, it only holds texture data and completion status.
- Fix an incorrect calculation in `calcLocalChunkCoordinate()`.
- Add a new function `calcLocalChunkOrigin()`.
- Refactor some class like `STPNearestNeighbourTextureBuffer` and heightfield renderer to make use of the chunk utility functions instead of calculating from scratch.
- Improve testing robustness for chunk calculations in `STPTest2DChunk`.