:dromedary_camel:Implement a regular language parser.
The regex library in C++ standard library is terrible, it is a good idea to implement our own.
Basically it is a regex parser, but it uses template as matching terms rather than regex string.
DFA/NFA are constructed at compile-time with no dynamic memory allocation in runtime.

The regular language parser `STPRegularLanguage`, resides in the algorithm+ package. Test is implemented as `STPTestRegLang.cpp`.

# Others
- Make sure all use of functions from header *cctype* have the input argument cast to *unsigned char* to avoid UB.